Offline refresh

+++ dedupCopiedFiles.py
@@ -46,6 +46,7 @@ from urllib.parse import unquote
 
 # Move to commmon
 def build_file_tree(items=[], recurse=False):
+  norecursedirs = ['__Trash']
   temp_dict = {}
   for item in args.items:
     if os.path.isfile(item):
@@ -54,76 +55,125 @@ def build_file_tree(items=[], recurse=False):
       item = os.getcwd()
     if os.path.isdir(item):
       if recurse:
+#        found = False
+#        for norecursedir in norecursedirs:
+#          if norecursedir in item:
+#            found = True
+# 
         for root, dirs, files in os.walk(item):
           for f in files:
-            temp_dict[os.path.abspath(os.path.join(root,f))] = item
+            found = False
+            for norecursedir in norecursedirs:
+              if norecursedir in item:
+                found = True
+            if not found:
+              temp_dict[os.path.abspath(os.path.join(root,f))] = item
       else:
         for possiblefile in os.listdir(item):
           if os.path.isfile(os.path.join(item, possiblefile)):
             temp_dict[os.path.abspath(os.path.join(item, possiblefile))] = item
   return sorted(temp_dict.keys())
 
-
-def delete_file(filename, trash_folder = None, actionfile = None):
-  if trash_folder:
-    basedir = os.path.dirname(filename)
-    trash_folder_pathed = os.path.join(basedir, trash_folder)
-    if os.path.isdir(trash_folder_pathed):
-      if not os.access(trash_folder_pathed, os.W_OK):
-        logger.error("Cannot write to trash folder %s" % trash_folder_pathed)
+# Move to commmon
+def delete_file(filename, trash_folder = None, actionfile = None, logger = None):
+  if logger:
+    logger.debug("In delete_file")
+  if os.path.isfile(filename): 
+    if trash_folder:
+      basedir = os.path.dirname(filename)
+      trash_folder_pathed = os.path.join(basedir, trash_folder)
+      if os.path.isdir(trash_folder_pathed):
+        if not os.access(trash_folder_pathed, os.W_OK):
+          if logger:
+            logger.error("Cannot write to trash folder %s" % trash_folder_pathed)
+          else:
+            print("Error: Cannot write to trash folder %s" % trash_folder_pathed)
+          sys.exit(1)
+      else:
+        try:
+          os.makedirs(trash_folder_pathed, exist_ok=True)
+        except OSError as exc: 
+          if exc.errno == errno.EEXIST and os.path.isdir(path):
+            pass
+          else:
+            if logger:
+              logger.error("Cannot create trash folder %s" % trash_folder_pathed)
+            else:
+              print("Error: Cannot create trash folder %s" % trash_folder_pathed)
+            sys.exit(1)
+        # Action statement
+        if actionfile:
+          actionfile.write("\n")
+          actionfile.write("Date: %s\n" % str(time.time()) )
+          actionfile.write("Action: CreateDir\n")
+          actionfile.write("Object: %s\n" % trash_folder_pathed)
+      existing_file = os.path.join(trash_folder_pathed, os.path.basename(filename))
+      if os.path.isfile(existing_file):
+        if logger:
+          logger.error("Would overwrite exsiting file in Trash folder. Please remove content in Trash folder.")
+        else:
+          print("Error: Would overwrite exsiting file in Trash folder. Please remove content in Trash folder.")
+        sys.exit(1)  
+      try:
+        shutil.move(filename,trash_folder_pathed)
+      except:
+        if logger:
+          logger.error("Cannot move file %s to trash folder %s" % (filename, trash_folder_pathed))
+        else:
+          print("Error: Cannot move file %s to trash folder %s" % (filename, trash_folder_pathed))
         sys.exit(1)
+      if actionfile:
+        actionfile.write("\n")
+        actionfile.write("Date: %s\n" % str(time.time()) )
+        actionfile.write("Action: MoveFile\n")
+        actionfile.write("Object: %s\n" % filename)
+        actionfile.write("Destination: %s\n" % existing_file)
     else:
       try:
-        os.mkdirs(trash_folder_pathed, exist_ok=True)
+        os.remove(filename)
       except:
-        logger.error("Cannot create trash folder %s" % trash_folder_pathed)
+        if logger:
+          logger.error("Cannot remove file %s" % filename) 
+        else:
+          print("Error: Cannot remove file %s" % filename)
         sys.exit(1)
-      # Action statement
-      actionfile.write("\n")
-      actionfile.write("Date: %s\n" % str(time.time()) )
-      actionfile.write("Action: CreateDir\n")
-      actionfile.write("Object: %s\n" % trash_folder_pathed)
-    existing_file = os.path.join(trash_folder_pathed, os.path.basename(filename))
-    if os.path.isfile(existing_file):
-      logger.error("Would overwrite exsiting file in Trash folder. Please remove content in Trash folder.")
-      sys.exit(1)  
-    try:
-      shutil.move(filename,trash_folder_pathed)
-    except:
-      logger.error("Cannot move file %s to trash folder %s" % (filename, trash_folder_pathed))
-      sys.exit(1)
-    actionfile.write("\n")
-    actionfile.write("Date: %s\n" % str(time.time()) )
-    actionfile.write("Action: MoveFile\n")
-    actionfile.write("Object: %s\n" % filename)
-    actionfile.write("Destination: %s\n" % existing_file)
+      if actionfile:
+        actionfile.write("\n")
+        actionfile.write("Date: %s\n" % str(time.time()) )
+        actionfile.write("Action: DeleteFile\n")
+        actionfile.write("Object: %s\n" % filename)
   else:
-    try:
-      os.remove(filename)
-    except:
-      logger.error("Cannot remove file %s" % filename) 
-      sys.exit(1)
-    actionfile.write("\n")
-    actionfile.write("Date: %s\n" % str(time.time()) )
-    actionfile.write("Action: DeleteFile\n")
-    actionfile.write("Object: %s\n" % filename)
+    if logger:
+      logger.error("File %s went away before we could delete" % filename)
+    else:
+      print("Error: File %s went away before we could delete" % filename)
        
-
-def move_file(fromfile, tofile, actionfile):
+# Move to commmon
+def move_file(fromfile, tofile, actionfile = None, logger = None):
+  if logger:
+    logger.debug("In move_file")
   if os.path.isfile(tofile):
-    logger.error("Move of %s would overwrite exsiting file %s." % (fromfile, tofile))
+    if logger:
+      logger.error("Move of %s would overwrite existing file %s." % (fromfile, tofile))
+    else:
+      print("Error: Move of %s would overwrite existing file %s." % (fromfile, tofile))
     sys.exit(1)
   else:
     try:
       shutil.move(fromfile,tofile)
-    except:
-      logger.error("Cannot move file %s to file %s" % (fromfile, tofile)) 
+    except Exception as e:
+      print(str(e))
+      if logger:
+        logger.error("Cannot move file %s to file %s" % (fromfile, tofile)) 
+      else:
+        print("Error: Cannot move file %s to file %s" % (fromfile, tofile))
       sys.exit(1)
-    actionfile.write("\n")
-    actionfile.write("Date: %s\n" % str(time.time()) )
-    actionfile.write("Action: MoveFile\n")
-    actionfile.write("Object: %s\n" % fromfile)
-    actionfile.write("Destination: %s\n" % tofile)
+    if actionfile:
+      actionfile.write("\n")
+      actionfile.write("Date: %s\n" % str(time.time()) )
+      actionfile.write("Action: MoveFile\n")
+      actionfile.write("Object: %s\n" % fromfile)
+      actionfile.write("Destination: %s\n" % tofile)
 
 
 # Local
@@ -151,7 +201,8 @@ parser.add_argument('-a', dest='aggressive', action='store_true', default=False,
 parser.add_argument('-noffend', dest='unixoffend', action='store_false', default=True, help='Remove typically offending Unix shell characters.')
 parser.add_argument('-R', dest='recurse', action='store_true', default=False, help='Recurse into directories.')
 parser.add_argument('-d', dest='debug', action='store_true', default=False, help='Enable debugging to standard out.')
-parser.add_argument('-report', dest='report', action='store_true', default=False, help='Report only, do not touch files.')
+parser.add_argument('-t', dest='report', action='store_true', default=False, help='Test and print report, do not touch files.')
+parser.add_argument('-v', dest='verbose', action='store_true', default=False, help='Verbose status.')
 parser.add_argument('-nlog', dest='actionlog', action='store_const', const=None, default='action.log', help='Disable action log generation.')
 parser.add_argument('-ntrash', dest='trash', action='store_const', const=None, default='__Trash',  help='Disable use of Trash folder.')
 parser.add_argument('items', nargs='+', default=None)
@@ -221,8 +272,15 @@ re_containsbonnie = re.compile('bonnie\+\+', re.IGNORECASE)
 
 
 checksum_cache = {}
+current_count = 0
 
 for item in files:
+  current_count += 1
+  if args.verbose:
+    print("%s/%s: %s" % (str(current_count), str(len(files)), item))
+  if not os.path.isfile(item):
+    logger.info("Skipping file %s since it was presumeably deleted or can't be read" % item)
+    break
   original = item
   filename = None
   suffix = None
@@ -294,6 +352,8 @@ for item in files:
 # 1F Filename contains
   filename = re_Id.sub('', filename)
   logger.debug("New File: %s" % filename)
+  filename = re_Id.sub('', filename)
+  logger.debug("New File: %s" % filename) # Run again due to nested .Id_
   filename = re_containscopyof.sub('', filename)
   logger.debug("New File: %s" % filename)
   filename = re_containscopyparenof.sub('', filename)
@@ -339,37 +399,40 @@ for item in files:
   deleted = False
   if args.aggressive:
     for localfile in localfiles:
-      (localfilename, localsuffix) = os.path.splitext(os.path.basename(localfile))
-      # If the local file is a possible varaint of the copy-stripped file
-      if filename in localfilename:
-        # Make sure we haven't seen ourselves
-        if not os.path.samefile(original, localfile):
-          if not localfile in checksum_cache.keys():
-            checksum_cache[localfile] = cksum(localfile)
-          # Did we find something that needs to be deleted
-          logger.debug("Localfile checksum: %s" % checksum_cache[localfile])
-          logger.debug("Original checksum: %s" % checksum_cache[original])   
-          if (checksum_cache[localfile] == checksum_cache[original]): # Same checksum so one has to go
-            # Do a Levenshtein distance to the copy-stripped file. Closet file to the stripped version is the best.
-            originalratio = int(fuzz.ratio(filename, original))
-            localratio = int(fuzz.ratio(filename, localfilename))
-            logger.debug("Localfile fuzzy ratio: %s" % str(localratio))
-            logger.debug("Original fuzzy ratio: %s" % str(originalratio))
-            if (originalratio > localratio):
-              if args.report:
-                print("Alternate file %s will be deleted due to hashing match with %s (aggressive)" % (localfile, original))
+      if not os.path.isfile(localfile):
+        (localfilename, localsuffix) = os.path.splitext(os.path.basename(localfile))
+        # If the local file is a possible varaint of the copy-stripped file
+        if filename in localfilename:
+          # Make sure we haven't seen ourselves
+          if not os.path.samefile(original, localfile):
+            if not localfile in checksum_cache.keys():
+              checksum_cache[localfile] = cksum(localfile)
+            # Did we find something that needs to be deleted
+            logger.debug("Localfile checksum: %s" % checksum_cache[localfile])
+            logger.debug("Original checksum: %s" % checksum_cache[original])   
+            if (checksum_cache[localfile] == checksum_cache[original]): # Same checksum so one has to go
+              # Do a Levenshtein distance to the copy-stripped file. Closet file to the stripped version is the best.
+              originalratio = int(fuzz.ratio(filename, original))
+              localratio = int(fuzz.ratio(filename, localfilename))
+              logger.debug("Localfile fuzzy ratio: %s" % str(localratio))
+              logger.debug("Original fuzzy ratio: %s" % str(originalratio))
+              if (originalratio > localratio):
+                if args.report:
+                  print("Alternate file %s will be deleted due to hashing match with %s (aggressive)" % (localfile, original))
+                else:
+                  delete_file(localfile, args.trash, af)
+#                  checksum_cache.pop(localfile)
+                  del checksum_cache[localfile]
+                continue
               else:
-                delete_file(localfile, args.trash, af)
-                checksum_cache.pop(localfile)
-              break
-            else:
+                if args.report:
+                  print("Duplicate file %s will be deleted due to hashing match with %s (aggressive)" % (original, localfile))
+                else:
+                  delete_file(original, args.trash, af)
+#                  checksum_cache.pop(original)
+                  del checksum_cache[original]
+                continue
               deleted = True
-              if args.report:
-                print("Duplicate file %s will be deleted due to hashing match with %s (aggressive)" % (original, localfile))
-              else:
-                delete_file(original, args.trash, af)
-                checksum_cache.pop(original)
-              break
 
   if not deleted:
 # 3 Postfixes -- not going to do any of these now, as I hope the logic above isn't so sloppy to need them
@@ -397,7 +460,7 @@ for item in files:
         if args.report:
           print("File %s will be moved from %s" % (newfile, original))
         else:
-          move_file(original, newfile, af)
+          move_file(original, newfile, af, logger)
           checksum_cache[newfile] = checksum_cache[original]
           checksum_cache.pop(original)
       else:
@@ -418,14 +481,24 @@ for item in files:
             checksum_cache.pop(original)
         else:
           # Would like to make this modular for but now we will use -(NNNN) as the sequence
-          for counter in range(10000):
+          rename_done = False
+          counter = 0
+          while not rename_done:
             if counter == 9999:
               logger.info("Counter overflow. This shouldn't happen")
               sys.exit(1)
             postscript = '-(' + format(counter, '04') + ')'
-            testfilename = filename + postscript
+            if filename:
+              testfilename = filename + postscript
+              testsuffix = suffix
+            elif suffix:
+              testfilename = filename
+              testsuffix = suffix + postscript
+            else:
+              logger.error("Error: We should have no filename and no suffix at the same time")
+              sys.exit(1)
             if suffix:
-              testrecombinedfile = testfilename + '.' + suffix
+              testrecombinedfile = testfilename + '.' + testsuffix
               newfile = os.path.join(basedir, testrecombinedfile)
               logger.debug("In counter loop with %s for %s" % (str(counter), newfile))
             else:
@@ -435,23 +508,27 @@ for item in files:
               if args.report:
                 print("Incremented file %s will be moved from %s" % (newfile, original))
               else:
-                move_file(original, newfile, af)
+                move_file(original, newfile, af, logger)
                 checksum_cache[newfile] = checksum_cache[original]
                 checksum_cache.pop(original)
-              break
+              rename_done = True
             else:
               if os.path.samefile(original, newfile):
-                logger.info("We shouldn't have got to this point in the final move, the original file shouldn't have gone into the loop")
-                sys.exit(1)
-              if not newfile in checksum_cache.keys():
+                logger.debug("We found ourselves in the rewrite, bail on rewriting %s under the assumption that the rewrite is the same file" % original)
+                rename_done = True
+               
+#                sys.exit(1)
+              elif not newfile in checksum_cache.keys():
                 checksum_cache[newfile] = cksum(newfile)
-              # Did we find something that needs to be deleted
-              logger.debug("Newfile (existing) checksum: %s" % checksum_cache[newfile])
-              logger.debug("Original checksum: %s" % checksum_cache[original])
-              if (checksum_cache[newfile] == checksum_cache[original]): # Same checksum so one has to go
-                if args.report:
-                  print("Duplicate file %s will be deleted due to hash map match on another file %s" % (original, newfile))
-                else:
-                  delete_file(original, args.trash, af)
-                  checksum_cache.pop(original)
-
+                # Did we find something that needs to be deleted
+                logger.debug("Newfile (existing) checksum: %s" % checksum_cache[newfile])
+                logger.debug("Original checksum: %s" % checksum_cache[original])
+                if (checksum_cache[newfile] == checksum_cache[original]): # Same checksum so one has to go
+                  if args.report:
+                    print("Duplicate file %s will be deleted due to hash map match on another file %s" % (original, newfile))
+                  else:
+                    delete_file(original, args.trash, af)
+                    checksum_cache.pop(original)
+                  rename_done = True
+            counter += 1 
+print("DONE.")
